UGENE-6862. The "Filter SAM/BAM" element doesn't save parameters values (#335)
* UGENE-6862. The "Filter SAM/BAM" element doesn't save parameters values

It is required that with each click on the checkbox, the data is
committed to the model. For this,
`ComboBoxWithChecksWidget::si_update(QWidget *)` was created. After the
signal connection, UGENE began to crash, so the code for creating the
model/view was refactored: `delete model` was removed.
Added GUI tests

* UGENE-6862. The "Filter SAM/BAM" element doesn't save parameters values
Fix GUI test

* UGENE-6862. The "Filter SAM/BAM" element doesn't save parameters values
Changes by review

* UGENE-6862. The "Filter SAM/BAM" element doesn't save parameters values
Changes by review 1
`si_update` isn't needed; data is committed by `si_valueChanged`

Co-authored-by: Evelina Biserova <biserova@unipro.ru>